// src/styles/types.ts
// import { Source } from 'react-native-fast-image'

export type DefaultTheme = {
  readonly PRIMARY_COLOR: string
  readonly SECONDARY_COLOR: string
  readonly PRIMARY_TEXT_COLOR: string
  readonly SECONDARY_TEXT_COLOR: string
  // readonly LOGO?: Source
  // readonly SECONDARY_LOGO?: Source
  readonly COMPANY_NAME: string
  readonly BODY_TEXT_COLOR: string
}